requirements.txt
yt-dlp
openai-whisper
google-generativeai
sqlalchemy
psycopg2-binary
google-api-python-client
requests